Fix SPSA momentum bug
Set `du_last` to the most recent `du` so that momentum is non-zero.

Yaoshu Xie <ysxie@iis.u-tokyo.ac.jp> pointed out that the momentum term is never actually used because `last_du` is never reset from its initial value of `0`. Thankfully this bug is equivalent to the original algorithm (i.e. no momentum) and nothing more, but thanks to Yaoshu Xie for catching it!